+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - `irma scheme verify` not detecting missing files in index 
 - Scheme verification/signing does not reject credentials with invalid revocation settings
+- Race conditions in database logic of revocation storage
 
 ### Added
 - E-mail address revalidation. At automated mailings (`irma keyshare tasks`) informing users their account is about to expire, (temporary) invalid e-mail addresses are withheld from processing for 5 days, preventing cluttering of the processing queue
@@ -22,6 +23,9 @@ As part of e-mail address revalidation:
 - MyIrma server: `/user` returns an additional field `revalidate_in_progress` in the JSON response body, indicating whether the e-mail address is being revalidated or not 
 - MyIrma server: `/user/delete` and `/email/remove` return a 400 status code if one or more e-mail addresses of the user are invalid
 - MyIrma server: `/email/remove` returns a 400 status code if one or more e-mail addresses of the user are invalid
+- Rename RevocationStorage's UpdateLatest function to LatestUpdates. This name better fits its behaviour. The functionality stays the same.
+- Validate revocation witness before revocation update is applied
+- RevocationStorage's EnableRevocation function does not return an error anymore if it has been enabled already
 
 ### Removed
 - Superfluous openssl package in Dockerfile